Renamed pslr_shared.h to pslr_api.h
