+++ libpktriggercord.h
@@ -36,7 +36,7 @@
 #ifndef LIBPKTRIGGERCORD_H
 #define LIBPKTRIGGERCORD_H
 
-#include "pslr_shared.h"
+#include "pslr_api.h"
 #include "pslr_log.h"
 #include "pslr.h"
 #include "pslr_lens.h"
