+++ pktriggercord-servermode.c
@@ -43,7 +43,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "pslr_shared.h"
 #include "pslr_log.h"
 #include "pslr.h"
 #include "pslr_lens.h"
